# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API for managing products and bundles",
        "title": "Product \u0026 Bundle Service API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/admin/bundles": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Bundles"
                ],
                "summary": "Create bundle",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Admin role",
                        "name": "Role",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "Bundle",
                        "name": "bundle",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_Sanuthiabey_CTSE-Product-and-Bundle-Service_internal_models.CreateBundleRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/github_com_Sanuthiabey_CTSE-Product-and-Bundle-Service_internal_models.Bundle"
                        }
                    }
                }
            }
        },
        "/admin/products": {
            "post": {
                "responses": {}
            }
        },
        "/admin/products/{id}": {
            "put": {
                "responses": {}
            },
            "delete": {
                "responses": {}
            }
        },
        "/admin/stock/reduce": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Stock"
                ],
                "summary": "Reduce stock",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Admin role",
                        "name": "Role",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "Stock reduction",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/api/bundles": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Bundles"
                ],
                "summary": "Get all bundles",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/github_com_Sanuthiabey_CTSE-Product-and-Bundle-Service_internal_models.Bundle"
                            }
                        }
                    }
                }
            }
        },
        "/api/products": {
            "get": {
                "description": "Retrieve list of products",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Products"
                ],
                "summary": "Get all products",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "limit",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "offset",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/github_com_Sanuthiabey_CTSE-Product-and-Bundle-Service_internal_models.Product"
                            }
                        }
                    }
                }
            }
        },
        "/api/products/{id}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Products"
                ],
                "summary": "Get product by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Product ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_Sanuthiabey_CTSE-Product-and-Bundle-Service_internal_models.Product"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/stock/validate": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Stock"
                ],
                "summary": "Validate stock",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "Stock request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "github_com_Sanuthiabey_CTSE-Product-and-Bundle-Service_internal_models.Bundle": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "mood": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "products": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_Sanuthiabey_CTSE-Product-and-Bundle-Service_internal_models.BundleProduct"
                    }
                }
            }
        },
        "github_com_Sanuthiabey_CTSE-Product-and-Bundle-Service_internal_models.BundleProduct": {
            "type": "object",
            "properties": {
                "product_id": {
                    "type": "string"
                },
                "quantity": {
                    "type": "integer"
                }
            }
        },
        "github_com_Sanuthiabey_CTSE-Product-and-Bundle-Service_internal_models.CreateBundleRequest": {
            "type": "object",
            "required": [
                "id",
                "name",
                "products"
            ],
            "properties": {
                "id": {
                    "type": "string"
                },
                "mood": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "products": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_Sanuthiabey_CTSE-Product-and-Bundle-Service_internal_models.BundleProduct"
                    }
                }
            }
        },
        "github_com_Sanuthiabey_CTSE-Product-and-Bundle-Service_internal_models.Product": {
            "type": "object",
            "properties": {
                "category": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "featured": {
                    "type": "boolean"
                },
                "id": {
                    "type": "string"
                },
                "image": {
                    "type": "string"
                },
                "mood": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "price": {
                    "type": "number"
                },
                "rating": {
                    "type": "number"
                },
                "stock": {
                    "type": "integer"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}